Sample from a causal model

Obtain samples from a causal model using the rejection sampling approach of Evans and Didelez (2024).

causalSamp(
  n,
  formulas = list(list(z ~ 1), list(x ~ z), list(y ~ x), list(~1)),
  pars,
  family,
  link = NULL,
  dat = NULL,
  method = "rejection",
  control = list(),
  seed
)

Arguments

n

number of samples required

formulas

list of lists of formulas

pars

list of lists of parameters

family

families for Z,X,Y and copula

link

list of link functions

dat

data frame of covariates

method

only "rejection" is valid

control

list of options for the algorithm

seed

random seed used for replication

Value

A data frame containing the simulated data.

Details

Samples from a given causal model using rejection sampling (or, if everything is discrete, direct sampling).

The entries for formula and family should each be a list with four entries, corresponding to the \(Z\), \(X\), \(Y\) and the copula. formula determines the model, so it is crucial that every variable to be simulated is represented there exactly once. Each entry of that list can either be a single formula, or a list of formulae. Each corresponding entry in family should be the same length as the list in formula or of length 1 (in which case it will be repeated for all the variables therein).

We use the following codes for different families of distributions: 0 or 5 = binary; 1 = normal; 2 = t-distribution; 3 = gamma; 4 = beta; 6 = log-normal.

The family variables for the copula are also numeric and taken from VineCopula. Use, for example, 1 for Gaussian, 2 for t, 3 for Clayton, 4 for Gumbel, 5 for Frank, 6 for Joe and 11 for FGM copulas.

pars should be a named list containing: either entries z, x, y and cop, or variable names that correspond to the LHS of formulae in formulas. Each of these should themselves be a list containing beta (a vector of regression parameters) and (possibly) phi, a dispersion parameter. For any discrete variable that is a treatment, you can also specify p, an initial proportion to simulate from (otherwise this defaults to 0.5).

Link functions for the Gaussian, t and Gamma distributions can be the identity, inverse or log functions. Gaussian and t-distributions default to the identity, and Gamma to the log link. For the Bernoulli the logit and probit links are available.

Control parameters are oversamp (default value 10), trace (default value 0, increasing to 1 increases verbosity of output), max_oversamp (default value 1000), warn (which currently does nothing), max_wt which is set to 1, and increases each time the function is recalled. Control parameters also include cop, which gives a keyword for the copula that defaults to "cop".

This function is kept largely for the replication of simulations from Evans and Didelez (2024).

References

Evans, R.J. and Didelez, V. Parameterizing and simulating from causal models (with discussion). Journal of the Royal Statistical Society, Series B, 2024.

Examples

pars <- list(z=list(beta=0, phi=1),
             x=list(beta=c(0,0.5), phi=1),
             y=list(beta=c(0,0.5), phi=0.5),
             cop=list(beta=1))
causalSamp(100, family = rep(1,4), pars = pars)
#>               z           x           y
#> 1   -0.22389510 -0.34787601  0.28618614
#> 2    2.61373434  2.87574031  1.66441281
#> 3    1.69278228  3.08328249  2.50716934
#> 4   -0.43664023 -0.13655094 -0.94279614
#> 5    0.13092897 -2.43231673 -1.72841194
#> 6    0.10197039 -1.00988912 -1.32951242
#> 7    1.40395294  0.13804381  0.33273691
#> 8   -0.12348178 -0.52132471 -0.06974845
#> 9    0.61199612 -0.54360206  1.07196971
#> 10  -1.01270697 -2.03434122 -1.23076313
#> 11   0.61336626  1.31104683  0.77956788
#> 12  -1.33808111 -1.54554151 -1.71040022
#> 13  -1.56447947 -0.67110798  0.94459729
#> 14   1.63476557  0.38731556  1.25563008
#> 15   0.46240514 -0.11720092  0.76599301
#> 16   0.74219097 -0.09089773 -0.21370773
#> 17  -0.01906428  1.69240485  1.60854483
#> 18   0.64273352  1.24363335  1.26735633
#> 19   1.12014147  0.86761579  0.77407409
#> 20  -0.04853157  0.13660255  0.37958191
#> 21   0.61440471 -1.78160171 -1.55318742
#> 22   1.23242342  1.37076362  1.39511676
#> 23  -0.11756081  0.05156592  0.40530960
#> 24   0.17125405  0.26684422  0.56231906
#> 25  -1.77754858 -1.72242579 -1.92248514
#> 26   2.46613947 -0.38645131  0.67186373
#> 27  -1.51648621 -0.79469659 -1.19258073
#> 28  -0.77989286 -0.03571429  0.27159260
#> 29   0.08937055 -0.49901567 -0.12993839
#> 30  -0.84349924  0.59651980 -0.62634199
#> 31   1.19240186  2.37262096  2.25729489
#> 32   0.13125732  0.34301794 -0.58911750
#> 33  -1.99346078 -0.57531299 -1.27728669
#> 34  -1.86626970 -1.19712361 -0.89214373
#> 35  -0.02897060 -0.84710930 -0.20502922
#> 36   0.02670126 -1.53730040  0.16231217
#> 37  -0.70485287 -0.21560573 -0.91573559
#> 38  -0.53648444 -2.04150075 -1.39984450
#> 39   0.61472474  0.49442126  0.98549521
#> 40  -1.05294254 -0.55691497  0.05164624
#> 41   1.81621157  0.63544246  1.04028025
#> 42   0.06953122 -0.40282643 -0.41717055
#> 43  -1.42184308  0.06229820 -0.15566914
#> 44   0.58314249  0.24282549  0.71561823
#> 45  -0.88946632  0.30375838  0.76692757
#> 46  -0.94068542 -0.53791479 -0.34988876
#> 47   0.65824943  0.15314322  0.20368821
#> 48   0.20549215  1.42121419  0.07752815
#> 49   1.22885614  1.81113530  1.00555471
#> 50   0.55920781  0.37453373 -0.53224766
#> 51  -0.53566658  0.37565473 -0.52244529
#> 52   0.01436555  1.08929930  0.72553165
#> 53  -0.20087976 -0.40486208 -0.59768587
#> 54  -1.27978984 -0.11141083 -0.10287384
#> 55  -2.21434258 -3.06115763 -1.67867940
#> 56  -1.37583913 -1.29342075 -1.15701971
#> 57  -0.05409516  1.81519476  2.06458937
#> 58  -1.35194763  0.08897170 -0.64404816
#> 59   0.87082276  1.83106910  0.99829460
#> 60  -0.53787073 -1.41267131 -1.33054822
#> 61  -0.93568502  0.32120105 -0.15419429
#> 62   1.54051150  1.13780937  0.30900187
#> 63   0.05085022  0.53431814 -0.16868018
#> 64  -0.56535675  0.31380818  0.29612190
#> 65  -0.46592345  0.12148165  0.42189647
#> 66   1.04126087 -0.11320663 -0.09310470
#> 67   0.04729857  0.59737582  1.43897280
#> 68  -1.05292594 -0.16183271  0.38610801
#> 69  -0.46313653 -1.38968197 -0.93002077
#> 70   0.88748695  0.05943483  0.16205558
#> 71  -1.81094381 -2.16870166 -2.82273379
#> 72  -0.40760200 -1.90688544 -0.49695357
#> 73  -1.73487979 -1.19572959 -1.21375189
#> 74  -0.06382091  1.01037730 -0.63076531
#> 75   0.01072759  1.65751510  0.94902097
#> 76  -0.91982712 -0.10330196  1.05162682
#> 77  -2.14685585 -0.65379298 -1.48941934
#> 78   0.72609978  1.61921664  1.26016429
#> 79  -0.39296603 -0.62610747  0.07879166
#> 80  -0.19583244  0.37604069 -0.35745809
#> 81   2.33301509  2.54267277  1.52743637
#> 82   0.41096707  2.00383112  0.98394765
#> 83   0.15517070  0.57355330  0.69018355
#> 84  -0.30098206  0.15723322  0.03866088
#> 85   0.86907803 -1.71434853 -1.42163686
#> 86  -0.99852518 -1.76219523 -1.99803828
#> 87  -1.66437531 -1.43828641 -0.76697160
#> 88  -1.78107814 -0.62765484 -1.53875180
#> 89   0.59955250  1.67981029  0.84941333
#> 90   1.11056415  0.26923731 -0.09902919
#> 91  -0.57462764 -1.52262372 -1.45092565
#> 92  -0.49217658 -0.23258588 -0.14357653
#> 93  -0.64776594 -1.03200669 -0.53573127
#> 94   1.78721182  1.30575440  0.88655842
#> 95  -0.80955310 -1.94702080 -1.04655728
#> 96  -1.81602341 -1.13577877 -1.22281318
#> 97  -1.40195317 -1.28985485 -2.37707246
#> 98  -0.63249137  0.65402334  0.36622634
#> 99  -0.65709747 -1.46495582 -0.38867511
#> 100 -1.48901635 -1.45567337 -1.99489868