Process formulas, families and parameters

Process formulas, families and parameters

process_inputs(
  formulas,
  family,
  pars,
  link,
  dat,
  kwd,
  method = "inversion",
  control = list()
)

process_formulas(formulas, len = 4)

process_family(family, dims, link = link, func_return = get_family)

Arguments

formulas

list of lists of formulas

family

families for variables and copula

pars

list of lists of parameters

link

list of link functions

dat

optional data frame of covariates

kwd

keyword for copula

method

either "inversion" (the default), "inversion_mv", or "rejection"

control

list of options for the algorithm

len

number of formulas

dims

number of variables in each class

func_return

function to use to process character arguments

Details

Function that processes and checks the validity of the main arguments used for simulating data. The control argument only requires values useful for obtaining the empirical (conditional) quantiles of pre-specified data.

For causl we use the get_family() function to process character based arguments, but we allow for other functions to be used in packages that build on this one.

Functions

process_formulas(): Process input for family variables

process_family(): Process input for family variables