Sample from a causal model

Obtain samples from a causal model parameterized as in Evans and Didelez (2024).

rfrugal(n, causl_model, control = list())

rfrugalParam(
  n,
  formulas = list(list(z ~ 1), list(x ~ z), list(y ~ x), list(~1)),
  family = c(1, 1, 1, 1),
  pars,
  link = NULL,
  dat = NULL,
  method = "inversion",
  control = list(),
  ...
)

Arguments

n

number of samples required

causl_model

object of class causl_model

control

list of options for the algorithm

formulas

list of lists of formulas

family

families for variables and copula

pars

list of lists of parameters

link

list of link functions

dat

optional data frame of covariates

method

either "inversion" (the default), "inversion_mv", or "rejection"

...

other arguments, such as custom families

Value

A data frame containing the simulated data.

Details

Samples from a given causal model under the frugal parameterization.

The entries for formula and family should each be a list with four entries, corresponding to the \(Z\), \(X\), \(Y\) and the copula. formula determines the model, so it is crucial that every variable to be simulated is represented there exactly once. Each entry of that list can either be a single formula, or a list of formulae. Each corresponding entry in family should be the same length as the list in formula or of length 1 (in which case it will be repeated for all the variables therein).

We use the following codes for different families of distributions:

val	family

0	binomial

1	gaussian

2	t

3	Gamma

4	beta

5	binomial

6	lognormal

11	ordinal

10	categorical

The family variables for the copula are also numeric and taken from VineCopula. Use, for example, 1 for Gaussian, 2 for t, 3 for Clayton, 4 for Gumbel, 5 for Frank, 6 for Joe and 11 for FGM copulas.

pars should be a named list containing variable names that correspond to the LHS of formulae in formulas. Each of these should themselves be a list containing beta (a vector of regression parameters) and (possibly) phi, a dispersion parameter. For any discrete variable that is a treatment, you can also specify p, an initial proportion to simulate from (otherwise this defaults to 0.5).

Link functions for the Gaussian, t and Gamma distributions can be the identity, inverse or log functions. Gaussian and t-distributions default to the identity, and Gamma to the log link. For the Bernoulli the logit, probit, and log links are available.

A variety of sampling methods are implemented. The inversion method with pair-copulas is the default (method="inversion"), but we cam also use a multivariate copula (method="inversion_mv") or even rejection sampling (method="rejection"). Note that the inveresion_mv method simulates the entire copula, so it cannot depend upon intermediate variables.

The only control parameters are cop: which gives a keyword for the copula that defaults to "cop"; quiet which defaults to FALSE but will reduce output if set to TRUE; and (if rejection sampling is selected) careful: this logical enables one to implement the full rejection sampling method, which means we do get exact samples (note this method is generally very slow, especially if we have an outlying value, so the default is FALSE).

Functions

rfrugalParam(): old function for simulation

Examples

pars <- list(z=list(beta=0, phi=1),
             x=list(beta=c(0,0.5), phi=1),
             y=list(beta=c(0,0.5), phi=0.5),
             cop=list(beta=1))
rfrugalParam(100, pars = pars)
#> Inversion method selected: using pair-copula parameterization
#>               z           x           y
#> 1    0.85040618 -0.61507469  0.99607640
#> 2    0.50272182 -1.24550454 -0.21134813
#> 3    0.09233139  0.96812656 -0.37216849
#> 4    0.10254351  0.89501636  0.52708162
#> 5    0.27283360  1.19170190  0.27463989
#> 6   -0.03707721  1.20327915  0.01815158
#> 7   -1.96390331 -2.09746524 -1.76601586
#> 8   -1.44599278 -1.93544135 -2.30979680
#> 9   -0.63758736 -0.79845182 -1.48911661
#> 10  -0.44369813  0.92775681  0.04583188
#> 11  -1.33950075 -1.11864813 -2.66379660
#> 12  -0.36993622 -0.41113158 -0.62026205
#> 13   0.26447602  0.60739876  0.21177828
#> 14  -1.57318815  0.61359346  0.26808965
#> 15  -0.39374270 -0.73818704 -0.26906397
#> 16  -1.86765164 -0.26689993 -0.28769790
#> 17  -0.10683542  0.89212107  1.14127552
#> 18  -0.15732709 -0.24437022  0.40186490
#> 19   0.76480415  1.40997853  0.61860617
#> 20   0.44841491  0.58917961  0.35752529
#> 21  -0.93486854  0.24850126 -0.37435561
#> 22   2.07863911 -0.16618607 -0.14962284
#> 23  -1.23314444 -1.90286628 -0.83824936
#> 24   1.91569335  0.92752221  0.97112896
#> 25  -1.77757719 -0.10326839  0.94045377
#> 26  -2.00895400 -0.95012117 -0.84695051
#> 27  -0.40803676 -0.14024030  0.27178988
#> 28   0.59230652 -0.32894790  0.09119383
#> 29  -1.63406986 -1.43916865 -0.51929060
#> 30   0.49645401 -1.25373350  0.13607438
#> 31  -0.22750054  0.64091038  0.97475825
#> 32   0.47965188  1.03282166  0.72261941
#> 33   0.31371103 -0.48192703  0.96208710
#> 34  -0.07049288 -1.69089375 -0.11029688
#> 35  -1.37434177 -0.89788578 -2.03978632
#> 36  -1.28319237 -0.71678606 -0.68154412
#> 37   0.66443839  0.14940312 -0.21440887
#> 38   0.18449306 -0.81787143  0.04778326
#> 39  -0.76383870 -0.71606383 -0.05267544
#> 40   0.26784589  0.58678703  0.65021180
#> 41   0.04400526 -1.52579053 -1.02805016
#> 42  -2.83886132 -0.77083218 -1.53559138
#> 43   1.27642651  1.92070673  1.01558941
#> 44   2.10614110  0.68711815  0.88691934
#> 45  -0.57464334  0.01260033  0.39559826
#> 46  -1.11798713  0.42933845 -1.01007120
#> 47  -0.35282569 -0.30500641 -0.85728742
#> 48   1.79084592  1.97379403  1.62764745
#> 49  -0.93612169 -1.09617551 -1.22859076
#> 50  -0.21350865  1.79352164  0.88239834
#> 51  -1.60327081 -1.41512069 -0.87842017
#> 52   1.40721257  0.24405640 -0.42658486
#> 53   0.42016720  0.49659479  0.67746506
#> 54   1.39988694  1.85260296  0.90002841
#> 55  -2.23888990 -1.14788569 -1.75547698
#> 56   0.22856861  3.27757892  2.26925408
#> 57  -0.44538298 -0.37543073 -0.90135972
#> 58  -0.47955025 -0.78619930 -1.37099243
#> 59  -2.38011795 -1.78276095 -1.13716213
#> 60  -1.56526006 -2.37461308 -1.45049647
#> 61  -1.69910415 -1.45724543 -2.22767991
#> 62  -0.78342534 -0.91252796 -0.48346520
#> 63  -0.82527872 -0.56960361 -0.81951313
#> 64  -0.62369161 -0.77924705 -0.20409409
#> 65   1.20826302  0.11988038  0.10153440
#> 66   1.10481124  0.59400213  1.92433041
#> 67   0.37802712  1.83416754  1.07389631
#> 68   0.35596653 -0.05701453  0.77761218
#> 69  -1.06909394  0.76137063  0.53514978
#> 70  -0.43690374 -0.33319685  0.19295472
#> 71  -1.09133694 -1.51190766 -0.90230940
#> 72  -1.30195784 -0.83162537 -0.93814016
#> 73  -1.79935218 -0.17569948 -0.59205577
#> 74  -1.21872783 -1.23578698 -2.14331611
#> 75  -0.45963736  0.55085154  0.79219784
#> 76   1.34315300  0.76079704 -0.25791560
#> 77   0.07053956  1.18170149  0.42192559
#> 78  -0.85452178  0.10094107  0.15733082
#> 79   1.92889916  1.34505016  1.39755538
#> 80  -0.27230028  0.40553390 -0.55814424
#> 81   1.00706570 -0.57558584  0.51585534
#> 82  -0.34465157 -0.37611251 -0.40800704
#> 83   0.27419560  0.50678927  0.50160603
#> 84   1.15486019  0.34568546 -0.19271635
#> 85  -2.32451258 -1.75232733 -2.36174802
#> 86  -1.19327556  1.46292569 -0.17223947
#> 87  -0.95682240  0.69162371  0.10322394
#> 88   0.33646305  0.20796451 -0.48436257
#> 89   1.24656003  1.53447453  1.49660445
#> 90   1.30459833  0.42312688  1.29783271
#> 91   1.29372874  1.06165506  0.52888032
#> 92   0.28431500 -1.17788190 -0.62789341
#> 93  -0.60652280  0.01066634 -1.00048927
#> 94   1.36897212 -0.25249610 -0.25018909
#> 95   1.45574586  0.46923468  1.52823361
#> 96  -0.74549215  1.04338019  0.73251355
#> 97  -0.10700916 -1.27466241  0.25849782
#> 98  -0.13400479 -0.17323458 -0.07896746
#> 99  -0.58746930 -0.64376219  0.93263817
#> 100  0.36371387  1.40562108  1.14981049